[common] Move IsAffine into symbolic_decompose
We do not want code in the basic scalartype-expression library to
depend on all of the polynomial analysis logic.

In the future, the BUILD labels for this will change, but the
function declaration and definition will stay where they are.

The co-authored-by line below refers to the moved code.
