+++ common/BUILD.bazel
@@ -194,6 +194,7 @@ drake_cc_library(
         "symbolic_chebyshev_polynomial.h",
         "symbolic_codegen.cc",
         "symbolic_codegen.h",
+        "symbolic_decompose.cc",
         "symbolic_environment.cc",
         "symbolic_environment.h",
         "symbolic_expression.cc",
@@ -232,6 +233,9 @@ drake_cc_library(
     ],
     hdrs = [
         "symbolic.h",
+        # TODO(#13833) We should split out this header into its own library,
+        # once we have the new subdirectory established.
+        "symbolic_decompose.h",
     ],
     deps = [
         ":bit_cast",
@@ -246,14 +250,11 @@ drake_cc_library(
     ],
 )
 
+# TODO(#13833) This library name is vestigial.  Once we are finished with the
+# new file layout in drake/common/symbolic/..., then we should deprecate this
+# library name.
 drake_cc_library(
     name = "symbolic_decompose",
-    srcs = [
-        "symbolic_decompose.cc",
-    ],
-    hdrs = [
-        "symbolic_decompose.h",
-    ],
     deps = [
         ":symbolic",
     ],
